[REFACTOR] : DTO 클래스와 Controller 클래스의 패키지 분리.
Controller의 개수가 늘어났으므로, DTO클래스와 Controller 클래스를 각각의 패키지로 분리해주었다.
