+++ be/sidedish-be/src/main/java/com/codesquad/sidedish/web/sidedish/DTO/ItemDTO.java
@@ -1,4 +1,4 @@
-package com.codesquad.sidedish.web.sidedish;
+package com.codesquad.sidedish.web.sidedish.DTO;
 
 import java.text.NumberFormat;
 import java.text.ParseException;
