+++ be/sidedish-be/src/main/java/com/codesquad/sidedish/web/sidedish/controller/SideSidedishesController.java
@@ -1,6 +1,7 @@
-package com.codesquad.sidedish.web.sidedish;
+package com.codesquad.sidedish.web.sidedish.controller;
 
 import com.codesquad.sidedish.utils.SampleDataFactory;
+import com.codesquad.sidedish.web.sidedish.DTO.ItemDTO;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
